backend/requirements.txt
aiohappyeyeballs==2.6.1
aiofiles==24.1.0
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
bcrypt==4.2.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-limiter==0.1.6
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.3.1
kombu==5.4.2
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-text-splitters==0.3.7
langsmith==0.3.18
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.65.5
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sentry-sdk==2.20.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.41.3
structlog==25.1.0
tenacity==9.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.1
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
anthropic
groq==0.31.0
beautifulsoup4==4.12.2
PyPDF2
pandas
filetype==1.2.0
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==0.6.23
html2text==2024.2.26
ebooklib==0.18
odfpy==1.4.1
striprtf==0.0.26
markdown==3.7
lxml==5.3.0
